+++ InternRegistrationForm/Controllers/TermController.cs
@@ -1,4 +1,5 @@
-﻿using InternRegistrationForm.Models;
+﻿using InternRegistrationForm.Classes;
+using InternRegistrationForm.Models;
 using InternRegistrationForm.Repositories;
 using Microsoft.AspNetCore.Mvc;
 
@@ -17,6 +18,11 @@ namespace InternRegistrationForm.Controllers
 
         public async Task<IActionResult> TermDates()
         {
+            if (HttpContext.Session.GetInt32("AdminId") == null)
+            {
+                return RedirectToRoute(nameof(HomeController), nameof(HomeController.LogIn));
+            }
+
             List<TermDatesModel> termDates = await _termDatesRepo.GetAll();
 
             return View(termDates);
@@ -24,6 +30,16 @@ namespace InternRegistrationForm.Controllers
 
         public IActionResult DeleteTerm(int id)
         {
+            if (HttpContext.Session.GetInt32("AdminId") == null)
+            {
+                return RedirectToRoute(nameof(HomeController), nameof(HomeController.LogIn));
+            }
+            if (HttpContext.Session.GetInt32("PermissionsLevel") == 0)
+            {
+                return RedirectToAction(nameof(HomeController.InvalidPermissions), NameOfController.ControllerName(nameof(HomeController)));
+            }
+
+
             _termDatesRepo.DeleteTerm(id);
             return RedirectToAction(nameof(TermDates));
         }
@@ -31,24 +47,64 @@ namespace InternRegistrationForm.Controllers
         [HttpGet]
         public IActionResult AddTerm()
         {
+            if (HttpContext.Session.GetInt32("AdminId") == null)
+            {
+                return RedirectToRoute(nameof(HomeController), nameof(HomeController.LogIn));
+            }
+
+            if (HttpContext.Session.GetInt32("PermissionsLevel") == 0)
+            {
+                return RedirectToAction(nameof(HomeController.InvalidPermissions), NameOfController.ControllerName(nameof(HomeController)));
+            }
+
             return View();
         }
 
         [HttpPost]
         public async Task<IActionResult> AddTerm(TermDatesModel term)
         {
+            if (HttpContext.Session.GetInt32("AdminId") == null)
+            {
+                return RedirectToRoute(nameof(HomeController), nameof(HomeController.LogIn));
+            }
+
+            if (HttpContext.Session.GetInt32("PermissionsLevel") == 0)
+            {
+                return RedirectToAction(nameof(HomeController.InvalidPermissions), NameOfController.ControllerName(nameof(HomeController)));
+            }
+
             await _termDatesRepo.AddTerm(term);
             return RedirectToAction(nameof(TermDates));
         }
 
         public IActionResult EditSingleTerm(TermDatesModel term)
         {
+            if (HttpContext.Session.GetInt32("AdminId") == null)
+            {
+                return RedirectToRoute(nameof(HomeController), nameof(HomeController.LogIn));
+            }
+
+            if (HttpContext.Session.GetInt32("PermissionsLevel") == 0)
+            {
+                return RedirectToAction(nameof(HomeController.InvalidPermissions), NameOfController.ControllerName(nameof(HomeController)));
+            }
+
             _termDatesRepo.EditDateForTrack(term);
             return RedirectToAction(nameof(TermDates));
         }
 
         public IActionResult EditSingleTermDate(int id)
         {
+            if (HttpContext.Session.GetInt32("AdminId") == null)
+            {
+                return RedirectToRoute(nameof(HomeController), nameof(HomeController.LogIn));
+            }
+
+            if (HttpContext.Session.GetInt32("PermissionsLevel") == 0)
+            {
+                return RedirectToAction(nameof(HomeController.InvalidPermissions), NameOfController.ControllerName(nameof(HomeController)));
+            }
+
             TermDatesModel termDatesModel = new TermDatesModel();
             termDatesModel.Id = id;
